+++ src/api/cameracontrol.hh
@@ -56,7 +56,7 @@ public:
      * \param[in] framewidth Width requested
      * \param[in] framheight Height requested
      */
-    bool preselect_framesize(size_t framewidth, size_t frameheight);
+    bool preselect_framesize(uint16_t framewidth, uint16_t frameheight);
 
     /**
      * Check if (any) cameradevice is available, acquire it if necessary.
@@ -116,7 +116,7 @@ public:
      * \param[out] bytesize total image width, i.e. in byte
      * \return False if no camera is opened or retrieving the values fails.
      */
-    bool get_properties(size_t& height, size_t& width, size_t& bytesize);
+    bool get_properties(uint16_t& height, uint16_t& width, size_t& bytesize);
 
     /**
      * Retrieves the frame properties from an opened device. No effect on
@@ -125,7 +125,7 @@ public:
      * \param[out] width (visible) image width, i.e. in pixel
      * \return False if no camera is opened or retrieving the values fails.
      */
-    bool get_resolution(size_t& width, size_t& height);
+    bool get_resolution(uint16_t& width, uint16_t& height);
 
     /**
      * Grab a frame if a camera is available.  Sets image_.
