+++ src/imaging/v4l2_camera.hh
@@ -132,7 +132,7 @@ public:
 
 protected:
     virtual bool retrieve_frame(tfv::Image& image);
-    virtual void retrieve_properties(size_t& width, size_t& height,
+    virtual void retrieve_properties(uint16_t& width, uint16_t& height,
                                      size_t& frame_bytesize);
     virtual void close(void);
